requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
black==21.11b1
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.9
django-crispy-forms==1.13.0
djangorestframework==3.12.4
idna==3.3
kombu==5.2.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
prompt-toolkit==3.0.23
pycountry==20.7.3
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
redis==4.0.2
regex==2021.11.10
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tomli==1.2.2
typing-extensions==4.0.0
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
